pyproject.toml
[project]
name = "toolprac"
version = "0.1.0"
description = "toolprac using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]>=0.134.0,<1.0.0",
    "google-api-python-client",
    "google-auth-httplib2",
    "google-auth-oauthlib"
]

[project.scripts]
toolprac = "toolprac.main:run"
run_crew = "toolprac.main:run"
train = "toolprac.main:train"
replay = "toolprac.main:replay"
test = "toolprac.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
